cephadm: move ContainerInfo class to container_engines.py
Move the ContainerInfo class, which has basically no dependencies, to
container engines module which is the current home for generic
*low-level* container things.
